+++ assets/js/third/woo/devs/wooQuantityButtons.js
@@ -17,7 +17,8 @@ WOOCOMMERCE QUANTITY BUTTONS
 ============================================== */
 function oceanwpWooQuantityButtons( $quantitySelector ) {
 
-	var $quantityBoxes;
+	var $quantityBoxes
+		$cart = $j( '.woocommerce div.product form.cart' );
 
 	if ( ! $quantitySelector ) {
 		$quantitySelector = '.qty';
@@ -42,7 +43,8 @@ function oceanwpWooQuantityButtons( $quantitySelector ) {
 
 		// Quantity input
 		if ( $j( 'body' ).hasClass( 'single-product' )
-			&& 'on' == oceanwpLocalize.floating_bar ) {
+			&& 'on' == oceanwpLocalize.floating_bar
+			&& ! $cart.hasClass( 'grouped_form' ) ) {
 			var $quantityInput = $j( '.woocommerce form input[type=number].qty' );
 			$quantityInput.on( 'keyup', function() { 
 				var qty_val = $j( this ).val();
@@ -59,7 +61,8 @@ function oceanwpWooQuantityButtons( $quantitySelector ) {
 
 			// If floating bar is enabled
 			if ( $j( 'body' ).hasClass( 'single-product' )
-				&& 'on' == oceanwpLocalize.floating_bar ) {
+				&& 'on' == oceanwpLocalize.floating_bar
+				&& ! $cart.hasClass( 'grouped_form' ) ) {
 				$quantityBox = $j( '.plus, .minus' ).closest( '.quantity' ).find( $quantitySelector );
 			} else {
 				$quantityBox = $j( this ).closest( '.quantity' ).find( $quantitySelector );