+++ inc/plugins/tgm-plugin-activation.php
@@ -24,18 +24,14 @@ function oceanwp_tgmpa_register() {
 			'force_activation'	=> false,
 		),
 		
+		array(
+			'name'				=> 'WPForms',
+			'slug'				=> 'wpforms-lite', 
+			'required'			=> false,
+			'force_activation'	=> false,
+		),
+		
 	);
-
-	// If OCean Pro Demos is not installed
-	if ( ! class_exists( 'Ocean_Pro_Demos' ) ) {
-
-		$plugins[]     = array(
-			'name'     => 'Ocean Demo Import',
-			'slug'     => 'ocean-demo-import',
-			'required' => false,
-		);
-
-	}
 
 	// If WooCommerce
 	if ( class_exists( 'WooCommerce' ) ) {