+++ partials/scroll-top.php
@@ -17,6 +17,10 @@ if ( ! oceanwp_display_scroll_up_button() ) {
 
 // Get arrow
 $arrow = apply_filters( 'ocean_scroll_top_arrow', get_theme_mod( 'ocean_scroll_top_arrow' ) );
-$arrow = $arrow ? $arrow : 'fa fa-angle-up'; ?>
+$arrow = $arrow ? $arrow : 'fa fa-angle-up'; 
 
-<a id="scroll-top" href="#"><span class="<?php echo esc_attr( $arrow ); ?>"></span></a>
+// Position
+$position = apply_filters( 'ocean_scroll_top_position', get_theme_mod( 'ocean_scroll_top_position' ) );
+$position = $position ? $position : 'right'; ?>
+
+<a id="scroll-top" class="scroll-top-<?php echo esc_attr( $position ); ?>" href="#"><span class="<?php echo esc_attr( $arrow ); ?>"></span></a>